Removed ViewController from iOS sample names and modified loading samples to correspond the change
