+++ src/iOS/Xamarin.iOS/ArcGISRuntime.Xamarin.Samples.iOS.csproj
@@ -209,18 +209,18 @@
     <Compile Include="Models\SampleModelExtensions.cs" />
     <Compile Include="Models\SampleStructureMap.cs" />
     <Compile Include="Models\SubCategoryModel.cs" />
-    <Compile Include="Samples\Layers\ArcGISMapImageLayerUrl\ArcGISMapImageLayerUrlViewController.cs" />
-    <Compile Include="Samples\Layers\ArcGISTiledLayerLocal\ArcGISTiledLayerLocalViewController.cs" />
-    <Compile Include="Samples\Layers\ArcGISTiledLayerUrl\ArcGISTiledLayerUrlViewController.cs" />
-    <Compile Include="Samples\Layers\ChangeSublayerVisibility\SublayersTableViewController.cs" />
-    <Compile Include="Samples\Map\AccessLoadStatus\AccessLoadStatusViewController.cs" />
-    <Compile Include="Samples\Map\ChangeBasemap\ChangeBasemapViewController.cs" />
-    <Compile Include="Samples\Map\ChangeViewpoint\ChangeViewpointViewController.cs" />
-    <Compile Include="Samples\Map\DisplayDrawingStatus\DisplayDrawingStatusViewController.cs" />
-    <Compile Include="Samples\Map\MapRotation\MapRotationViewController.cs" />
-    <Compile Include="Samples\Map\OpenExistingMap\OpenExistingMapViewController.cs" />
-    <Compile Include="Samples\Map\SetMinMaxScale\SetMinMaxScaleViewController.cs" />
-    <Compile Include="Samples\Symbology\SimpleMarkerSymbol\SimpleMarkerSymbolViewController.cs" />
+    <Compile Include="Samples\Layers\ArcGISMapImageLayerUrl\ArcGISMapImageLayerUrl.cs" />
+    <Compile Include="Samples\Layers\ArcGISTiledLayerLocal\ArcGISTiledLayerLocal.cs" />
+    <Compile Include="Samples\Layers\ArcGISTiledLayerUrl\ArcGISTiledLayerUrl.cs" />
+    <Compile Include="Samples\Layers\ChangeSublayerVisibility\SublayersTable.cs" />
+    <Compile Include="Samples\Map\AccessLoadStatus\AccessLoadStatus.cs" />
+    <Compile Include="Samples\Map\ChangeBasemap\ChangeBasemap.cs" />
+    <Compile Include="Samples\Map\ChangeViewpoint\ChangeViewpoint.cs" />
+    <Compile Include="Samples\Map\DisplayDrawingStatus\DisplayDrawingStatus.cs" />
+    <Compile Include="Samples\Map\MapRotation\MapRotation.cs" />
+    <Compile Include="Samples\Map\OpenExistingMap\OpenExistingMap.cs" />
+    <Compile Include="Samples\Map\SetMinMaxScale\SetMinMaxScale.cs" />
+    <Compile Include="Samples\Symbology\SimpleMarkerSymbol\SimpleMarkerSymbol.cs" />
     <Compile Include="ViewControllers\SamplesViewController.cs" />
     <Compile Include="ViewControllers\CategoriesViewController.cs" />
     <Compile Include="ViewControllers\CategoriesViewController.designer.cs">
@@ -228,9 +228,9 @@
     </Compile>
     <Compile Include="ViewControllers\SearchResultsUpdater.cs" />
     <Compile Include="ViewControllers\SearchResultsViewController.cs" />
-    <Compile Include="Samples\Layers\ChangeSublayerVisibility\ChangeSublayerVisibilityViewController.cs" />
-    <Compile Include="Samples\Layers\ChangeSublayerVisibility\ChangeSublayerVisibilityViewController.designer.cs">
-      <DependentUpon>ChangeSublayerVisibilityViewController.cs</DependentUpon>
+    <Compile Include="Samples\Layers\ChangeSublayerVisibility\ChangeSublayerVisibility.cs" />
+    <Compile Include="Samples\Layers\ChangeSublayerVisibility\ChangeSublayerVisibility.designer.cs">
+      <DependentUpon>ChangeSublayerVisibility.cs</DependentUpon>
     </Compile>
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.iOS.CSharp.targets" />
